small bugfix

+++ fetch_authorities.py
@@ -20,19 +20,20 @@ try:
                 if i < 6:
                     i = i + 1
                     print("fetching record " + str(i))
-                    auth_id = re.search('.*\/authorities\/.*\/(.*)', result.link).group(1)
-                    auth_file_url = 'http://lccn.loc.gov/' + auth_id + '/marcxml'
-                    time.sleep(20)
-                    urllib.urlretrieve (auth_file_url, 'tmp.marcxml')
-                    try:
-                        records = marcxml.parse_xml_to_array('tmp.marcxml')
-                        writer = MARCWriter(file('authorities_to_load.mrc', 'ab'))
-                        for r in records:
-                            writer.write(r)
-                            writer.close()
-                        os.remove('tmp.marcxml')
-                    except SAXParseException:
-                        print('Could not parse MARCXML file')
+                    if 'authorities' in result.link:
+                        auth_id = re.search('.*\/authorities\/.*\/(.*)', result.link).group(1)
+                        auth_file_url = 'http://lccn.loc.gov/' + auth_id + '/marcxml'
+                        time.sleep(20)
+                        urllib.urlretrieve (auth_file_url, 'tmp.marcxml')
+                        try:
+                            records = marcxml.parse_xml_to_array('tmp.marcxml')
+                            writer = MARCWriter(file('authorities_to_load.mrc', 'ab'))
+                            for r in records:
+                                writer.write(r)
+                                writer.close()
+                            os.remove('tmp.marcxml')
+                        except SAXParseException:
+                            print('Could not parse MARCXML file')
                 else:
                     break
 except IOError:
